requirements.txt
google-cloud-vision
opencv-python~=4.8.1.78
requests~=2.31.0